Add navbar code to just front page
In the future, we'll have to copy and paste the navbar HTML and JS onto
every new page we have. Also we'll need to update all of their HTML's
in the addition of a new tab. So it's probably better to batch that
as much as possible.

Also some addition content on the front page.